Module nids2::game

Module to hold game-system functions and static variables

Structs

BASE_COLOR_DISABLED

BASE_COLOR_FOCUSED

BASE_COLOR_NORMAL

BASE_COLOR_PRESSED

BORDER_COLOR_DISABLED

BORDER_COLOR_FOCUSED

BORDER_COLOR_NORMAL

BORDER_COLOR_PRESSED

BORDER_WIDTH

LOADED_TEXTURES

Hash Map of object types that are loaded into memory. Each map element contains an Arc with a texture and an ObjectConfig object. It is static so that all modules which import game can use it.

ObjectConfig

Holds the data for an object type. Note that this cannot create an instance of an object; it just holds the default configuration and spritesheet information.

TEXT_COLOR_DISABLED

Functions

color_init

Initializes all the default color pallette colors.

destroy

Clear the memory that has been allocated for textures and object configurations. Essentially forces the game to forget all object types.

init

Initializes the game memory. Should only be done once during the program. Call destroy() to clear the memory created by this. After destroying the game, it is possible to instantiate it again.

mutex_get

Extract a copy of something from a mutex.

mutex_set

Sets the value in the given mutex to be equivalent or a copy of val