Struct nids2::object::GenericObject

pub struct GenericObject {
    pub obj_id: i32,
    pub pos: Position,
    pub depth: i32,
    pub side: i32,
    pub side_index: i32,
    pub object_data: Arc<(Texture2D, ObjectConfig)>,
    pub side_shift_speed: i32,
    pub b_box: Option<Rectangle>,
    pub depthmod: i32,
    pub colormod: Color,
    // some fields omitted
}

The base for all objects. Grabs data from the LOADED_TEXTURES static variable and uses it to initialize an object of a known type.

Fields
obj_id: i32pos: Positiondepth: i32side: i32side_index: i32object_data: Arc<(Texture2D, ObjectConfig)>side_shift_speed: i32b_box: Option<Rectangle>depthmod: i32colormod: Color
Implementations

impl GenericObject

pub fn new(id: i32, obj_type: i32, pos: Option<Position>) -> Self

Create a new instance of an object of given type with a unique ID. The position the object is created in is either given by the user or is (0,0).

pub fn set_side(&mut self, side: u32) -> &mut Self

Set which sprite side the object is using

pub fn set_index(&mut self, index: u32) -> &mut Self

Set which sprite subimage to display

pub fn set_shift(&mut self, shift: i32) -> &mut Self

Set the frequency with which the sprite subimage is incremented. Lower values increment faster and 0 does not increment.

pub fn inc_index(&mut self) -> &mut Self

Increment the sprite subimage, wrapping back to 0 if it reaches the last subimage.

pub fn dec_index(&mut self) -> &mut Self

Decrement the sprite subimage, wrapping to the last subimage if it attempts to go before the fist subimage.

pub fn inc_side(&mut self) -> &mut Self

Increment the current sprite side, wrapping back around if it reaches the last side.

pub fn dec_side(&mut self) -> &mut Self

Decrement the current sprite side, wrapping to the end if it reaches the zeroth side.

pub fn move_x(&mut self, amt: i32, max: i32) -> &mut Self

Move the object by a given amount in the x direction, not reaching the max value. Has range [0, max)

pub fn move_x_unchecked(&mut self, amt: i32) -> &mut Self

pub fn move_y(&mut self, amt: i32, max: i32) -> &mut Self

Move the object by a given amount in the y direction, not reaching the max value. Has range [0, max)

pub fn move_y_unchecked(&mut self, amt: i32) -> &mut Self

pub fn pos_aligned(&self, grid: i32) -> bool

Check if the object’s position is aligned with a given grid.

pub fn get_name(&self) -> String

Return the name of the object’s type.

pub fn get_side(&self) -> i32

pub fn get_index(&self) -> i32

pub fn get_default_shift(&self) -> i32

pub fn update_depth(&mut self) -> &mut Self

pub fn width(&self) -> i32

pub fn height(&self) -> i32