Trait nids2::save::Saveable

pub trait Saveable<T> where
    T: Saveable<T>, {
    fn to_bytes(&self) -> Vec<u8>;

    fn from_bytes(bytes: &[u8]) -> Result<SaveInfo<T>, Box<dyn Error>>;
}

Trait converting from a type to a byte array so that it can be written to a file on the disk. Prefer big-endian endianess for consistency. from_bytes can return a different type than Self. That type is T. This is used for certain types, such as str and Ref which cannot themselves be created and returned, but which can be saved to bytes.

Example

impl Saveable<str> for str { /*...*/ } // wouldn't work because `from_bytes` cannot return str as it has no definite size.
impl Saveable<String> for str { /*...*/ } // would work because `String` can be returned.

Required methods

fn to_bytes(&self) -> Vec<u8>

To big-endian byte vector

fn from_bytes(bytes: &[u8]) -> Result<SaveInfo<T>, Box<dyn Error>>

From big-endian byte vector to type T. T is not necessarily the same as Self.