Module nids2::util

Module to hold utility code. Mostly drawing functions, but includes some functions to mess with data, objects, and other useful utils.

Functions

advanced_input

Creates a safe text-input box which respects backspaces, shifts, and most regular characters. Must be given a keyboardKey reference because RaylibDrawHandle doesn’t implement DerefMut

draw_text_centered

Print text using the default font with given fontsize. Text is centered on x and y.

ds_draw_slider_centered

Draw a slider width a given width and height, centered w.r.t. X and Y axes and With the title text above it.

ds_draw_toggle_rounded

Draw a toggle with the default style. Toggle has rounded corners (as opposed to RGui’s toggle)

ds_draw_toggle_rounded_centered

Draw a toggle with the default style, rounded corners, and centered at the position given by the rec’s x and y.

ds_rounded_button

Draws a rounded button with the default rgui style Returns a boolean for whether it was pressed and a vector2 which points to the top left corner of the drawn button.

ds_rounded_button_centered

Draws a rounded button centered. Returns a boolean for whether it was pressed and a vector2 which points to the top left corner of the drawn button.

ds_rounded_rectangle

Draws a rounded rectangle with the default rgui style

ds_rounded_rectangle_lines

Draws a rounded rectangle outline with the default rgui style

ds_scroll_selection

A Scrollable selection box with mouse control

ds_scroll_selection_auto

A scroll selection box that automatically determines the right height to draw itself to display a given number of items. If the given number of items is invalid, it will display 1 item.

ds_scroll_selection_ex

A Scrollable selection box with extended configuration

get_all_objects

Get Vector of all object types’ info.

get_all_objects_sorted

Get Hashmap with all object types sorted by their categories.

get_all_types

Get Vector of all unique categories that contain objects.

max

rect_midpoint

Get centerpoint of a rectangle

rel_rect

Adds x and y to rec.x and rec.y.